+++ packages/vue-component-library/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@ucla-library-monorepo/ucla-library-website-components-v1.81.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library-monorepo/ucla-library-website-components-v1.80.1...@ucla-library-monorepo/ucla-library-website-components-v1.81.0) (2026-08-05)
+
+
+### Features
+
+* LADI-5240-3_add-external-article-entry-type ([#968](https://github.com/UCLALibrary/ucla-library-website-components/issues/968)) ([67b62d7](https://github.com/UCLALibrary/ucla-library-website-components/commit/67b62d7a2ad48da7b2dc14e4f08b1f70f0f60116))
+
 # [@ucla-library-monorepo/ucla-library-website-components-v1.80.1](https://github.com/UCLALibrary/ucla-library-website-components/compare/@ucla-library-monorepo/ucla-library-website-components-v1.80.0...@ucla-library-monorepo/ucla-library-website-components-v1.80.1) (2026-08-01)